+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.0] - 2023-06-08
+
 ### Added
 
 - Helicopter RPM indicator with configurable limits
@@ -198,7 +200,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release with `demo` and `rpctask` sources
 
 <!-- prettier-ignore -->
-[Unreleased]: https://gitlab.com/Maarrk/lidia/-/compare/v0.8.1...dev
+[Unreleased]: https://gitlab.com/Maarrk/lidia/-/compare/v0.9.0...dev
+[0.9.0]: https://gitlab.com/Maarrk/lidia/-/compare/v0.8.1...v0.9.0
 [0.8.1]: https://gitlab.com/Maarrk/lidia/-/compare/v0.8.0...v0.8.1
 [0.8.0]: https://gitlab.com/Maarrk/lidia/-/compare/v0.7.1...v0.8.0
 [0.7.1]: https://gitlab.com/Maarrk/lidia/-/compare/v0.7.0...v0.7.1